backend/requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.1.1
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
zope.interface==8.0.1
